Overhaul organization of output functionality (again)
- close #177

- SOILWAT2-standalone and STEPWAT2 produce output in form of text (csv) files
- rSOILWAT2 and STEPWAT2 produce output in form of (in-memory) arrays

All functionality is shared across repositories and cannot be moved to a specific one; thus, I keep all output functionality in SOILWAT2.

- text-based output is organized in a new module `SW_Output_outtext`
- in-memory array-based output is organized in a new module `SW_Output_outarray`

- renamed file `SW_Output_SOILWAT2.c` to `SW_Output_get_functions.c`
- all `get_XXX` functionality from `rSOILWAT2` merged into here (`STEPWAT2` functionality is work in progress)

- SOILWAT2-standalone passes all unit tests and produces identical output as before
